refactor ui to have full ui handling in App struct
