+++ src/main.rs
@@ -14,7 +14,7 @@ struct Args {
 }
 
 #[tokio::main]
-async fn main() -> Result<(), String> {
+async fn main() -> Result<(), Box<dyn std::error::Error>> {
     let args = Args::parse();
 
     let config = config::init(&args.config_path)?;
@@ -28,11 +28,13 @@ async fn main() -> Result<(), String> {
 
     let requester = backend::nc_request::NCRequest::new(&config).expect("cannot create NCRequest");
 
-    match backend::nc_talk::NCTalk::new(requester, &config).await {
-        Ok(backend) => ui::run(backend, &config).await.expect("crashed"),
+    let backend = match backend::nc_talk::NCTalk::new(requester, &config).await {
+        Ok(backend) => backend,
         Err(why) => {
-            log::error!("Failed to create backend because: {why}");
+            panic!("Failed to create backend because: {}", why);
         }
     };
-    Ok(())
+    let mut ui: ui::app::App<'_, _> = ui::app::App::new(backend, &config);
+
+    ui.run(&config).await
 }
